refactor(api): move update referer route to team context
