+++ api/lib/routes.js
@@ -2,7 +2,6 @@ import * as healthcheck from '../src/shared/application/healthcheck/index.js';
 import * as authentication from './application/authentication/index.js';
 import * as campaignParticipations from './application/campaign-participations/index.js';
 import * as certificationCenterMemberships from './application/certification-center-memberships/index.js';
-import * as certificationCenters from './application/certification-centers/index.js';
 import * as frameworks from './application/frameworks/index.js';
 import * as organizations from './application/organizations/index.js';
 import * as scoOrganizationLearners from './application/sco-organization-learners/index.js';
@@ -11,7 +10,6 @@ import * as users from './application/users/index.js';
 const routes = [
   authentication,
   campaignParticipations,
-  certificationCenters,
   certificationCenterMemberships,
   healthcheck,
   organizations,
